feat(data): a field narrower than a byte is a channel like any other
`RawForm` said whole bytes only, so every one-bit flag and 3-bit selector an
ODIS project describes was dropped: 175,606 channels project-wide, 155,426 of
them single bits. On the reference car's fifteen units it is 2,732 channels,
taking what the tool can say from 3,963 of 6,886 to 6,695 — 97% of what the
file describes.

The within-byte bit order is established, not assumed. ASAM MCD-2D counts
BIT-POSITION from the least significant bit, and the car's own answers agree
twice over, from the parked and driving dumps in research/dumps:

  710/2A49 engine_running reads 0 parked and 1 driving counted from the LSB,
  and 0 both times counted from the MSB.

  7E0/20A4 carries eight flags named `cylinder 1` … `cylinder 8`. Counted from
  the LSB, cylinders 1-4 change between the dumps and 5-8 stay zero; counted
  from the MSB it is the other way round — on an engine with four cylinders.

A field that crosses a byte boundary is still refused. It needs a rule for
which end the bits continue from and nothing here has evidence for one; 97% do
not cross, so refusing the rest beats guessing at it.

The channel key becomes a bit offset rather than a byte offset, for the reason
the key exists at all: one byte holds eight flags, and a key counted in bytes
would give all eight of them one identity — the same loss one level down.

+++ crates/vagcan/src/watch/favourites.rs
@@ -78,8 +78,8 @@ pub fn parse(text: &str) -> BTreeSet<Key> {
 /// tool and read back by it, and guessing a control unit would silently move
 /// somebody's mark elsewhere.
 ///
-/// **The third part is the byte the field starts at, and a file written before
-/// it existed has two parts.** Those read as byte 0, which is what they meant:
+/// **The third part is the bit the field starts at, and a file written before
+/// it existed has two parts.** Those read as bit 0, which is what they meant:
 /// back then a channel was an identifier, and the only field anything could
 /// show was the one at the start of the response.
 fn parse_key(key: &str) -> Option<Key> {
@@ -87,7 +87,7 @@ fn parse_key(key: &str) -> Option<Key> {
 	let request = u16::from_str_radix(parts.next()?.trim(), 16).ok()?;
 	let did = u16::from_str_radix(parts.next()?.trim(), 16).ok()?;
 	let offset = match parts.next() {
-		Some(text) => u8::from_str_radix(text.trim(), 16).ok()?,
+		Some(text) => u32::from_str_radix(text.trim(), 16).ok()?,
 		None => 0,
 	};
 	// A fourth part is not a key this tool ever wrote.
